feat: add version structure list support
In order to support multi-client support. The version structure list
need to be persistent in the disk. So this commit first adds RPC support
for the server.

When performing any file operation, client should first download the
version structure list. When handling, it should update the version
structure list.

In this commit, there is no efficiency improvement, it aims at
correctness.